application run through 1 file
