+++ ADMIN_API_BP_LE/config.py
@@ -1,0 +1,5 @@
+from pymongo import MongoClient
+
+client = MongoClient("localhost", 27017)
+db = client["Slotzz"]
+collection = db["Account_holders"]
